shell: parent should close read end of pipe, child write end
This will finally make processes not hang (due to open FDs
pointing to pipe ends). However, it only works for a single
pipe. For multiple pipes need to do the same thing, just for
all other pipes (i.e. a generalized version of this commit).

Will address this in a future commit.